requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.0.1
pkg-resources==0.0.0
pytz==2018.9
xmltodict==0.12.0
